Refactor cart file structure
